Changes to a hybrid SDOM/Wags model (#9)
This version of Deku takes ideas from @paf31's SDOM as well as conversations with Yugiohxlight and blends them together into something that is as fast as the original while being more ergonomic and almost 0 typeclasses. The main motivation for the change was being able to port Deku to JS/TS in order to integrate it with a `three.js` event loop.