Hit a stumbling block.
For this to work, it needs to know that it's out of scope.
That means that, in addition to checking if the id is present, we _also_
need to check if the current scope is valid for the portal.

To do this, we'll need to redo the scoping mechanism to concatenate scopes
and then check for belonging.